+++ client/components/authentication/CreateUser.js
@@ -1,0 +1,1 @@
+//ability to create new user. this would need hash and bcrypt to encrypt the password
